Ajustes logica CRUD

+++ src/main/java/io/github/monthalcantara/estudojpa/domain/Cidade.java
@@ -22,7 +22,7 @@ public class Cidade implements Serializable{
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	
 	private String nome;
 	
@@ -30,7 +30,7 @@ public class Cidade implements Serializable{
 	@JoinColumn(name = "estado_id")
 	private Estado estado;
 	
-	public Cidade(Long id, String nome, Estado estado) {
+	public Cidade(Integer id, String nome, Estado estado) {
 
 		super();
 		this.id = id;
@@ -42,11 +42,11 @@ public class Cidade implements Serializable{
 	}
 
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
